Changes to the error of the state machine and fixes to data sent
it doesn't send the time yet, changes made to be able to test in the
computer the quaternions equations and functioning. The ERROR state in
the state machine now sets the state to init or configuration depending
on where it came from. There are some failures in long time readings,
need to look into that.